backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via
    #   django
    #   djoser
astor==0.8.1
    # via flake8-simplify
attrs==21.2.0
    # via
    #   flake8-bugbear
    #   flake8-eradicate
    #   flake8-multiline-containers
    #   pytest
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.4
    # via requests
click==8.0.1
    # via pip-tools
cognitive-complexity==1.2.0
    # via flake8-cognitive-complexity
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
coverage==5.5
    # via
    #   -r requirements.in
    #   pytest-cov
cryptography==3.4.7
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-extra-fields
django-cors-headers==3.8.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-split-settings==1.1.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-extra-fields
djangorestframework-simplejwt==4.8.0
    # via djoser
djoser==2.1.0
    # via -r requirements.in
drf-extra-fields==3.1.1
    # via -r requirements.in
eradicate==2.0.0
    # via flake8-eradicate
faker==8.12.0
    # via mixer
flake8==3.9.2
    # via
    #   flake8-absolute-import
    #   flake8-bugbear
    #   flake8-commas
    #   flake8-django
    #   flake8-eradicate
    #   flake8-isort
    #   flake8-multiline-containers
    #   flake8-mutable
    #   flake8-pep3101
    #   flake8-print
    #   flake8-printf-formatting
    #   flake8-pytest
    #   flake8-quotes
    #   flake8-simplify
    #   flake8-walrus
flake8-absolute-import==1.0
    # via -r requirements.in
flake8-bugbear==21.4.3
    # via -r requirements.in
flake8-cognitive-complexity==0.1.0
    # via -r requirements.in
flake8-commas==2.0.0
    # via -r requirements.in
flake8-django==1.1.2
    # via -r requirements.in
flake8-eradicate==1.1.0
    # via -r requirements.in
flake8-fixme==1.1.1
    # via -r requirements.in
flake8-isort==4.0.0
    # via -r requirements.in
flake8-mock==0.3
    # via -r requirements.in
flake8-multiline-containers==0.0.18
    # via -r requirements.in
flake8-mutable==1.2.0
    # via -r requirements.in
flake8-pep3101==1.3.0
    # via -r requirements.in
flake8-pie==0.14.0
    # via -r requirements.in
flake8-plugin-utils==1.3.2
    # via flake8-pytest-style
flake8-print==4.0.0
    # via -r requirements.in
flake8-printf-formatting==1.1.2
    # via -r requirements.in
flake8-pytest==1.3
    # via -r requirements.in
flake8-pytest-style==1.5.0
    # via -r requirements.in
flake8-quotes==3.2.0
    # via -r requirements.in
flake8-simplify==0.14.1
    # via -r requirements.in
flake8-todo==0.7
    # via -r requirements.in
flake8-walrus==1.1.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.2
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.9.3
    # via flake8-isort
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mixer==7.1.2
    # via -r requirements.in
nano==0.10.0
    # via -r requirements.in
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.0
    # via pytest
pep517==0.11.0
    # via pip-tools
pillow==8.3.1
    # via
    #   -r requirements.in
    #   reportlab
pip-tools==6.2.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
psycopg2-binary==2.9.1
    # via -r requirements.in
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-print
    #   flake8-todo
pycparser==2.20
    # via cffi
pyflakes==2.3.1
    # via flake8
pyjwt==2.1.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==2.12.1
    # via -r requirements.in
pytest-django==4.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.1
    # via django
reportlab==3.6.1
    # via -r requirements.in
requests==2.26.0
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via social-auth-core
six==1.16.0
    # via
    #   flake8-print
    #   python-dateutil
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.1
    # via django
testfixtures==6.18.1
    # via flake8-isort
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
tomli==1.2.1
    # via pep517
typing-extensions==3.10.0.0
    # via flake8-pie
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.6
    # via requests
wheel==0.37.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
